+++ lib/blocs/infographics/InfoGraphicsListBloc.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:pikobar_flutter/configs/SharedPreferences/LabelNew.dart';
 import 'package:pikobar_flutter/constants/collections.dart';
+import 'package:pikobar_flutter/models/LabelNewModel.dart';
 import 'package:pikobar_flutter/repositories/InfoGraphicsRepository.dart';
 import './Bloc.dart';
 
@@ -9,6 +11,16 @@ class InfoGraphicsListBloc
     extends Bloc<InfoGraphicsListEvent, InfoGraphicsListState> {
   final InfoGraphicsRepository _repository = InfoGraphicsRepository();
   StreamSubscription _subscription;
+  DateTime currentDay = DateTime(
+      DateTime.now().year, DateTime.now().month, DateTime.now().day - 2);
+  DateTime yesterday = DateTime(
+      DateTime.now().year, DateTime.now().month, DateTime.now().day - 3);
+
+  // DateTime currentDay = DateTime.now();
+  //   // DateTime yesterday = DateTime(
+  //   //     DateTime.now().year, DateTime.now().month, DateTime.now().day - 1);
+
+  List<LabelNewModel> dataLabel = [];
 
   InfoGraphicsListBloc() : super(InitialInfoGraphicsListState());
 
@@ -28,6 +40,11 @@ class InfoGraphicsListBloc
   Stream<InfoGraphicsListState> _mapLoadInfoGraphicsListToState(
       {int limit, String infoGraphicsCollection}) async* {
     yield InfoGraphicsListLoading();
+    String label = await LabelNewSharedPreference.getLabelNewInfoGraphics();
+    if (label != null) {
+      dataLabel = LabelNewModel.decode(label);
+    }
+    print('panjang datalabel ' + dataLabel.length.toString());
     _subscription?.cancel();
     _subscription = infoGraphicsCollection == kAllInfographics
         ? _repository.getAllInfographicList().listen((event) {
@@ -37,19 +54,62 @@ class InfoGraphicsListBloc
             });
             dataListAllinfographics.sort(
                 (b, a) => a['published_date'].compareTo(b['published_date']));
+            _insertDataLabel(dataListAllinfographics);
             add(InfoGraphicsListUpdate(dataListAllinfographics));
           })
         : _repository
             .getInfoGraphics(
                 limit: limit, infoGraphicsCollection: infoGraphicsCollection)
             .listen(
-              (data) => add(InfoGraphicsListUpdate(data)),
-            );
+            (data) {
+              _insertDataLabel(data);
+              add(InfoGraphicsListUpdate(data));
+            },
+          );
   }
 
   Stream<InfoGraphicsListState> _mapUpdateInfoGraphicListToState(
       InfoGraphicsListUpdate event) async* {
     yield InfoGraphicsListLoaded(event.infoGraphicsList);
+  }
+
+  _insertDataLabel(List<DocumentSnapshot> listData) async {
+    listData.forEach((dataInfographic) {
+      var dataDate = DateTime.fromMillisecondsSinceEpoch(
+          dataInfographic['published_date'].seconds * 1000);
+      if (dataDate.day == currentDay.day &&
+              dataDate.month == currentDay.month ||
+          dataDate.day == yesterday.day && dataDate.month == yesterday.month) {
+        print('print masuk sini  ' +
+            DateTime.fromMillisecondsSinceEpoch(
+                    dataInfographic['published_date'].seconds * 1000)
+                .toString() +
+            "hari " +
+            DateTime.fromMillisecondsSinceEpoch(
+                    dataInfographic['published_date'].seconds * 1000)
+                .day
+                .toString() +
+            'judul ' +
+            dataInfographic['title'].toString());
+        LabelNewModel labelNewModel =
+            LabelNewModel(id: dataInfographic.id.toString(), isRead: '0');
+
+        print('cekk isinya dong ' +
+            dataLabel.contains(labelNewModel.id).toString());
+
+        var data = dataLabel
+            .where((test) => test.id.toLowerCase().contains(labelNewModel.id));
+
+        if (data.isEmpty) {
+          dataLabel.add(labelNewModel);
+        }
+      }
+    });
+
+    print('cekk isi dataLabel  = ' + dataLabel.length.toString());
+
+    String encodeData = LabelNewModel.encode(dataLabel);
+    await LabelNewSharedPreference.setLabelNewInfoGraphics(encodeData);
   }
 
   @override
